+++ src/interfaces/yt/context.ts
@@ -465,8 +465,18 @@ export interface VideoActionsMenuRenderer {
 }
 
 export interface TopLevelButton {
+  segmentedLikeDislikeButtonRenderer?: SegmentedLikeDislikeButtonRenderer;
   toggleButtonRenderer?: TopLevelButtonToggleButtonRenderer;
   buttonRenderer?: TopLevelButtonButtonRenderer;
+}
+
+export interface SegmentedLikeDislikeButtonRenderer {
+  likeButton: YTLikeButton;
+  dislikeButton: YTLikeButton;
+}
+
+export interface YTLikeButton {
+  toggleButtonRenderer: TopLevelButtonToggleButtonRenderer;
 }
 
 export interface TopLevelButtonButtonRenderer {
@@ -728,7 +738,7 @@ export interface VideoOwnerRenderer {
   thumbnail: Background;
   title: Byline;
   navigationEndpoint: VideoOwnerRendererNavigationEndpoint;
-  subscriberCountText: YTRunContainer;
+  subscriberCountText: YTText;
   trackingParams: string;
   badges: MetadataBadgeRendererContainer[];
   membershipButton: MembershipButton;
@@ -1125,7 +1135,57 @@ export interface ContinuationCommand {
   request: string;
 }
 
-export interface FrameworkUpdates {}
+export interface FrameworkUpdates {
+  entityBatchUpdate: EntityBatchUpdate;
+}
+
+export interface EntityBatchUpdate {
+  mutations: EntityMutation[];
+  timestamp: UpdateTimestamp;
+}
+
+export interface UpdateTimestamp {
+  seconds: string;
+  nanos: number;
+}
+
+export interface EntityMutation {
+  entityKey: string;
+  type: string;
+  payload: EntityMutationPayload;
+}
+
+export interface EntityMutationPayload {
+  emojiFountainDataEntity?: EmojiFountainDataEntity;
+  booleanEntity?: BooleanEntity;
+}
+
+export interface EmojiFountainDataEntity {
+  key: string;
+  reactionBuckets: ReactionBucket[];
+  updateTimeUsec: string;
+}
+
+export interface ReactionBucket {
+  reactions?: Reaction[];
+  totalReactions: number;
+  duration: BucketDuration;
+  intensityScore?: number;
+}
+
+export interface Reaction {
+  key: string;
+  value: number;
+}
+
+export interface BucketDuration {
+  seconds: string;
+}
+
+export interface BooleanEntity {
+  key: string;
+  value: boolean;
+}
 
 export interface OnResponseReceivedEndpoint {
   clickTrackingParams: string;